+++ src/main/java/wf3/project/alpha_betise/repositories/EvenementRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import wf3.project.alpha_betise.entities.Evenement;
 
 @Repository
-public interface EvenementRepository extends JpaRepository<Evenement, Integer> {
+public interface EvenementRepository extends JpaRepository<Evenement, Long> {
 
 }
