+++ src/main/java/wf3/project/alpha_betise/services/CommentaireUtilisateurService.java
@@ -1,15 +1,16 @@
 package wf3.project.alpha_betise.services;
 
 import java.util.List;
+
 import wf3.project.alpha_betise.entities.CommentaireUtilisateur;
 
 public interface CommentaireUtilisateurService {
 
 	public List<CommentaireUtilisateur> getAll();
 
-	public CommentaireUtilisateur get(Integer id) throws Exception;
+	public CommentaireUtilisateur get(Long id) throws Exception;
 
 	public CommentaireUtilisateur post(CommentaireUtilisateur commentaireUtilisateur);
 
-	public void delete(Integer id);
+	public void delete(Long id);
 }
