+++ src/main/java/wf3/project/alpha_betise/services/EtatStockService.java
@@ -8,9 +8,9 @@ public interface EtatStockService {
 
 	public List<EtatStock> getAll();
 
-	public EtatStock get(Integer id) throws Exception;
+	public EtatStock get(Long id) throws Exception;
 
 	public EtatStock post(EtatStock etatStock);
 
-	public void delete(Integer id);
+	public void delete(Long id);
 }
